Move grants to traits
